Crypto/requirements.txt
numpy==1.26.4
pandas==2.2.1
google-cloud-bigquery==3.19.0
scikit-learn==1.4.1.post1
shap==0.45.0
xgboost==2.0.3
python-dotenv
hyperliquid-python-sdk

